pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-agent-ot"
version = "0.1.0"
description = "Multi-Agent Context Processor for OT data (MQTT, OPC UA, Modbus)"
authors = [{ name="Your Name", email="your@email.com" }]
dependencies = [
    "redis",
    "fastapi",
    "uvicorn",
    "sentence-transformers",
    "faiss-cpu",
    "paho-mqtt",
    "opcua",
    "pymodbus",
    "pyyaml"
]

[project.scripts]
mcpai = "mcpai.main:main"
